pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.0.0a4",
    "langchain-anthropic>=0.3.20",
    "sqlmodel>=0.0.24",
]

[dependency-groups]
dev = [
    "langsmith>=0.4.27",
    "python-dotenv>=1.1.1",
    "ruff>=0.13.0",
]

[tool.ruff]
line-length = 100
